+++ agent-control/src/agent_type.rs
@@ -13,5 +13,6 @@ pub mod runtime_config;
 pub mod templates;
 pub mod templates_function;
 pub mod trivial_value;
+pub mod validation;
 pub mod variable;
 pub mod version_config;